Crate axipi

ArceOS Inter-Processor Interrupt (IPI) primitives.

Structs

Callback

A callback function that will be called when an [IpiEvent] is received and handled.

MulticastCallback

A Callback that can be called multiple times. It’s used for multicast IPI events.

Functions

init

Initialize the per-CPU IPI event queue.

ipi_handler

The handler for IPI events. It retrieves the events from the queue and calls the corresponding callbacks.

run_on_cpu

Executes a callback on the specified destination CPU via IPI.

run_on_each_cpu

Executes a callback on all other CPUs via IPI.